mesh: migrate CSR neighbor lists to 64-bit offsets / total_pairs
neighbor_list_t and gpu_neighbor_list_t stored int total_pairs +
int *offsets, ceiling 2.15e9 pairs. m11i + TURB_DIFF_DYNAMIC measured
1.97e9 (92% of the int CSR ceiling); a larger problem overflows into
negative offsets -> illegal device addresses in the compact pass.

- total_pairs: int -> int64_t; offsets: int* -> int64_t* (CSR row
  pointers = cumulative pair counts). neighbors stays int*: its values
  are particle indices bounded by num_total < 2^31 -- only the array
  length is 64-bit. num_active stays int.
- Value-narrowing sweep (compiler does NOT catch these): prefix-scan
  accumulator + offsets writeback; every device + host CSR-walk loop
  index (start/end/nn/idx/dst) across gpu_neighbor_list, the runner,
  density_gpu, and all *_gpu.cc / host consumers; CPU builders'
  total_pairs accumulators.
- gradient_evaluate_gpu / hydro_evaluate_gpu signatures take int64_t
  csr offsets + total_pairs (extern decls in 3 callers updated).
- Allocation byte counts cast through size_t after the >=0 guard.
  Retired the total_ll > INT_MAX abort; replaced with a 64-bit sanity
  guard (negative => scan corruption; oversized => alloc overflow).
- Lazy-drift dirty-mark: guard the one >INT_MAX call site by escalating
  to a full-pool mark rather than truncating the index count.
- ~12 printf/PRINT_STATUS %d of total_pairs -> %lld.
- NGL debug dump magic v1/v2 -> v3/v4 (offsets now int64, 8 B/entry);
  tools/ngl_diff.py + ngl_forensic.py updated to match.

Mac (Kokkos OpenMP) + Vista (CUDA) compile clean.

+++ galaxy_sf/radfb_local.cc
@@ -331,7 +331,7 @@ static void hii_gpu_path(const std::vector<HIISourcePrep>& src,
          * below indexes it from host code; deep_copy once to a host buffer.
          * (Host memcpy from CudaSpace segfaults on GH200.) */
         if(gnl.total_pairs > 0) {
-            gnl_neighbors_host.resize(gnl.total_pairs);
+            gnl_neighbors_host.resize((size_t)gnl.total_pairs);
             gpu_ngb_copy_neighbors_to_host(&gnl, gnl_neighbors_host.data());
         }
     }
@@ -345,12 +345,14 @@ static void hii_gpu_path(const std::vector<HIISourcePrep>& src,
     std::vector<int> cand_buf;
     std::vector<int> ngb_scratch;
     for(int aa = 0; aa < num_src; aa++) {
-        int nl_start = gnl.offsets[aa], nl_end = gnl.offsets[aa+1];
-        int nl_n = nl_end - nl_start;
+        /* offsets are int64_t (CSR row pointers); per-source candidate count
+         * is bounded by num_total < 2^31 so the local int nl_n is safe. */
+        int64_t nl_start = gnl.offsets[aa], nl_end = gnl.offsets[aa+1];
+        int nl_n = (int)(nl_end - nl_start);
         /* Filter ghosts (singledomain). The helper rechecks Type/Mass/radius. */
         cand_buf.clear();
         if((int)cand_buf.capacity() < nl_n) cand_buf.reserve(nl_n > 0 ? nl_n : 1);
-        for(int nn = nl_start; nn < nl_end; nn++) {
+        for(int64_t nn = nl_start; nn < nl_end; nn++) {
             int j_cand = gnl_neighbors[nn];
             if(j_cand >= local_count) continue; /* skip ghosts */
             cand_buf.push_back(j_cand);
